+++ src/molecules/ElectricityDelivered/index.js
@@ -8,33 +8,27 @@ import LoadingSpinner from '../../atoms/LoadingSpinner';
 import { useQuery, gql, NetworkStatus } from '@apollo/client';
 import { FormattedNumber } from 'react-intl';
 
-import { DateTime } from 'luxon';
+import { ELECTRICITY_EXCHANGE_OVER_TIME } from '../../fragments';
 
-import { ELECTRICITY_EXCHANGE } from '../../fragments';
-
-const ELECTRICITY_DELIVERED = gql`
-	${ELECTRICITY_EXCHANGE}
-	query electricityDelivered($start: Int! $end: Int! $unit: ElectricEnergyOverTimeUnit) {
-		electricityExchange(start: $start end: $end) {
-			...ElectricityExchangeFields
+const TODAY_ELECTRICITY_DELIVERED = gql`
+	${ELECTRICITY_EXCHANGE_OVER_TIME}
+	query todayElectricityExchangeReceived($unit: ElectricEnergyOverTimeUnit) {
+		todayElectricityExchange {
+			...ElectricityExchangeOverTimeFields
 		}
 	}
 `;
 
-export default function ElectricityDelivered({ start, end }) {
-	const { error, data, networkStatus } = useQuery(ELECTRICITY_DELIVERED, {
+export default function ElectricityDelivered() {
+	const { error, data, networkStatus } = useQuery(TODAY_ELECTRICITY_DELIVERED, {
 		pollInterval: 60000 * 5,
 		notifyOnNetworkStatusChange: true,
-		variables: {
-			start: start || Math.round(DateTime.now().startOf('day').toSeconds()),
-			end: end || Math.floor(DateTime.now().endOf('day').toSeconds())
-		}
 	});
 
 	if (error) return <Message type="error">{error.message}</Message>;
 	return (
 		<div className={styles.container}>
-			{networkStatus !== NetworkStatus.loading ? data && data.electricityExchange ? <span><FormattedNumber value={data.electricityExchange.delivered} /> kWh</span> : 0 + ' kWh' : <Skeleton width="3em" />}
+			{networkStatus !== NetworkStatus.loading ? data && data.todayElectricityExchange ? <span><FormattedNumber value={data.todayElectricityExchange.delivered} /> kWh</span> : 0 + ' kWh' : <Skeleton width="3em" />}
 			<LoadingSpinner isHidden={networkStatus !== NetworkStatus.poll} diameter="16px" borderWidth="3px" />
 		</div>
 	);