+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.36.0] - 2022-10-22
+### Changed
+- Better queries for day usage widgets.
+
+### Fixed
+- Side nav active highlight for dashboard.
+- Dashboard widget usage current loading state.
+
 ## [1.35.1] - 2022-10-18
 ### Fixed
 - Dashboard widget loading state.